requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter
dvc
joblib
pyyaml
mlflow
xgboost
lightgbm
optuna
shap
plotly
dash
streamlit
flask
pytest
black
flake8